CDD-1555: Phase 1 of test clean up
- Rename and move acceptance tests to component
- Move common expected results to a single trait
- tidied up some of the test wording
- TODO
    - deduplicate unit tests
    - deduplicate unhappy path component tests and put in unit tests
    - pair with QA to check if component test are covered in their automated tests and remove from their tests
    - pair with QA and create acceptance test repo and start writing acceptance tests (or use their existing ones)
