+++ test/component/CustomsDeclarationCancellationSpec.scala
@@ -14,22 +14,24 @@
  * limitations under the License.
  */
 
-package acceptance
+package component
 
 import com.github.tomakehurst.wiremock.client.WireMock.{postRequestedFor, urlEqualTo, verify}
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, Matchers, OptionValues}
+import play.api.libs.json.{JsObject, JsString}
 import play.api.mvc._
 import play.api.test.FakeRequest
-import play.api.test.Helpers._
+import play.api.test.Helpers.{status, _}
 import uk.gov.hmrc.customs.declaration.model.{ApiSubscriptionKey, VersionOne, VersionTwo}
 import util.FakeRequests._
+import util.RequestHeaders.X_CONVERSATION_ID_NAME
 import util.TestData._
 import util.externalservices.{ApiSubscriptionFieldsService, AuthService, MdgWcoDecService}
-import util.{CustomsDeclarationsExternalServicesConfig, RequestHeaders}
+import util.{AuditService, CustomsDeclarationsExternalServicesConfig, RequestHeaders}
 
 import scala.concurrent.Future
 
-class CustomsDeclarationCancellationSpec extends AcceptanceTestSpec
+class CustomsDeclarationCancellationSpec extends ComponentTestSpec with AuditService with ExpectedTestResponses
   with Matchers
   with OptionValues
   with BeforeAndAfterAll
@@ -44,13 +46,6 @@ class CustomsDeclarationCancellationSpec extends AcceptanceTestSpec
     ApiSubscriptionKey(clientId = clientId, context = "customs%2Fdeclarations", VersionOne)
 
   private val apiSubscriptionKeyForXClientIdV2 = apiSubscriptionKeyForXClientIdV1.copy(version = VersionTwo)
-  private val UnauthorisedError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>UNAUTHORIZED</code>
-      |  <message>Unauthorised request</message>
-      |</errorResponse>
-    """.stripMargin
 
   override protected def beforeAll() {
     startMockServer()
@@ -132,10 +127,10 @@ class CustomsDeclarationCancellationSpec extends AcceptanceTestSpec
       Then("a response with a 401 (UNAUTHORIZED) status is received")
       status(result) shouldBe UNAUTHORIZED
 
-      And("the response body is empty")
-      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedError)
+      And("the response body is Unauthorised error")
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
 
-      And("the request was authorised with AuthService")
+      And("the AuthService was called")
       verifyAuthServiceCalledForCsp()
     }
 
@@ -158,7 +153,7 @@ class CustomsDeclarationCancellationSpec extends AcceptanceTestSpec
       And("the response body is empty")
       contentAsString(result) shouldBe 'empty
 
-      And("the request was authorised with AuthService")
+      And("the AuthService was called")
       verifyAuthServiceCalledForNonCsp()
     }
 
@@ -177,8 +172,8 @@ class CustomsDeclarationCancellationSpec extends AcceptanceTestSpec
       Then("a response with a 401 (UNAUTHORIZED) status is received")
       status(result) shouldBe UNAUTHORIZED
 
-      And("the response body is empty")
-      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedError)
+      And("the response body is unauthorised Error")
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
 
       And("the request was authorised with AuthService")
       verifyAuthServiceCalledForNonCsp()
@@ -293,6 +288,247 @@ class CustomsDeclarationCancellationSpec extends AcceptanceTestSpec
       And("the api-subscription-fields service was called with value of X-Client-ID header and api version 2.0")
       verifyGetSubscriptionFieldsNotCalled(apiSubscriptionKeyForXClientIdV1)
       verifyGetSubscriptionFieldsCalled(apiSubsKey = apiSubscriptionKeyForXClientIdV2)
+    }
+  }
+
+  //***  UNHAPPY PATH SPECS ***
+
+  feature("The API handles errors as expected") {
+
+    scenario("Response status 400 when user submits an xml payload that does not adhere to schema (and would be valid for submission)") {
+      Given("the API is available")
+      stubAuditService()
+      authServiceAuthorizesCSP()
+      val request = ValidSubmissionRequest.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"invalid xml\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestErrorIncorrectEndpoint)
+    }
+
+    scenario("Response status 400 when user submits an xml payload that does not adhere to schema") {
+      Given("the API is available")
+      stubAuditService()
+      authServiceAuthorizesCSP()
+      val request = InvalidCancellationRequest.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"invalid xml\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestError)
+    }
+
+    scenario("Response status 400 when user submits an xml payload that does not adhere to schema having multiple errors") {
+      Given("the API is available")
+      stubAuditService()
+      authServiceAuthorizesCSP()
+      val request = InvalidCancellationRequestWith3Errors.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"invalid xml\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestErrorWith2Errors)
+    }
+
+    scenario("Response status 400 when user submits a malformed xml payload") {
+      Given("the API is available")
+      stubAuditService()
+      authServiceAuthorizesCSP()
+      val request = MalformedXmlRequest.fromCsp.copyFakeRequest(method = POST, uri = endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"malformed xml body\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(MalformedXmlBodyError)
+    }
+
+    scenario("Response status 400 when user submits a non-xml payload") {
+      Given("the API is available")
+      val request = ValidCancellationRequest.fromCsp
+        .withJsonBody(JsObject(Seq("something" -> JsString("I am a json"))))
+        .copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"malformed xml body\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(MalformedXmlBodyError)
+    }
+
+    scenario("Response status 500 when user submits a request without any client id headers") {
+      Given("the API is available")
+      val request = NoClientIdIdHeaderCancellationRequest.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 500 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe INTERNAL_SERVER_ERROR
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"Internal server error\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InternalServerError)
+    }
+
+    scenario("Response status 406 when user submits a request without Accept header") {
+      Given("the API is available")
+      val request = NoAcceptHeaderCancellationRequest.copyFakeRequest(method = POST, uri = endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 406 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe NOT_ACCEPTABLE
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"invalid Accept header\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InvalidAcceptHeaderError)
+    }
+
+    scenario("Response status 406 when user submits a request with an invalid Accept header") {
+      Given("the API is available")
+      val request = InvalidAcceptHeaderCancellationRequest.copyFakeRequest(method = POST, uri = endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 406 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe NOT_ACCEPTABLE
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"invalid Accept header\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InvalidAcceptHeaderError)
+    }
+
+    scenario("Response status 415 when user submits a request with an invalid Content-Type header") {
+      Given("the API is available")
+      val request = InvalidContentTypeHeaderCancellationRequest.copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 415 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe UNSUPPORTED_MEDIA_TYPE
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"invalid Content-Type header\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InvalidContentTypeHeaderError)
+    }
+
+    scenario("Response status 400 when a CSP user submits a request without a X-Badge-Identifier header") {
+      Given("the API is available")
+      val request = InvalidCancellationRequestWithoutXBadgeIdentifier.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"Bad Request\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestErrorXBatchIdentifierMissingOrInvalid)
+    }
+
+    scenario("Response status 400 when a CSP user submits a request with an invalid X-Badge-Identifier header") {
+      Given("the API is available")
+      val request = InvalidCancellationRequestWithInvalidXBadgeIdentifier.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"Bad Request\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestErrorXBatchIdentifierMissingOrInvalid)
+    }
+
+    scenario("Response status 500 when user submits a valid request but downstream call to DMS fails with an HTTP error") {
+
+      Given("the API is available")
+      val request = ValidCancellationRequest.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      setupMdgWcoDecServiceToReturn(status = NOT_FOUND)
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 500 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe INTERNAL_SERVER_ERROR
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"Internal server error\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InternalServerError)
     }
   }
 }
