+++ test/component/CustomsDeclarationClearanceSpec.scala
@@ -14,39 +14,40 @@
  * limitations under the License.
  */
 
-package acceptance
+package component
 
-import org.scalatest.{Matchers, OptionValues}
+import com.github.tomakehurst.wiremock.client.WireMock.{postRequestedFor, urlEqualTo, verify}
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, Matchers, OptionValues}
 import play.api.libs.json.{JsObject, JsString}
 import play.api.mvc._
-import play.api.test.Helpers._
-import util.AuditService
+import play.api.test.FakeRequest
+import play.api.test.Helpers.{status, _}
+import uk.gov.hmrc.customs.declaration.model.{ApiSubscriptionKey, VersionOne, VersionTwo}
 import util.FakeRequests._
 import util.RequestHeaders.X_CONVERSATION_ID_NAME
-import util.externalservices.{AuthService, MdgWcoDecService}
+import util.TestData._
+import util.externalservices.{ApiSubscriptionFieldsService, AuthService, MdgWcoDecService}
+import util.{AuditService, CustomsDeclarationsExternalServicesConfig, RequestHeaders}
 
 import scala.concurrent.Future
 
-class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
-  with Matchers with OptionValues with AuthService with MdgWcoDecService with AuditService {
+class CustomsDeclarationClearanceSpec extends ComponentTestSpec with AuditService with ExpectedTestResponses
+  with Matchers
+  with OptionValues
+  with BeforeAndAfterAll
+  with BeforeAndAfterEach
+  with MdgWcoDecService
+  with ApiSubscriptionFieldsService
+  with AuthService {
 
   private val endpoint = "/clearance"
 
-  private val BadRequestError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>BAD_REQUEST</code>
-      |  <message>Payload is not valid according to schema</message>
-      |  <errors>
-      |     <error>
-      |       <code>xml_validation_error</code>
-      |       <message>cvc-complex-type.3.2.2: Attribute 'foo' is not allowed to appear in element 'Declaration'.</message>
-      |     </error>
-      |  </errors>
-      |</errorResponse>
-    """.stripMargin
+  private val apiSubscriptionKeyForXClientIdV1 =
+    ApiSubscriptionKey(clientId = clientId, context = "customs%2Fdeclarations", version = VersionOne)
 
-  private val BadRequestErrorWith2Errors =
+  private val apiSubscriptionKeyForXClientIdV2 = apiSubscriptionKeyForXClientIdV1.copy(version = VersionTwo)
+
+  protected override val BadRequestErrorWith2Errors =
     """<?xml version="1.0" encoding="UTF-8"?>
       |<errorResponse>
       |  <code>BAD_REQUEST</code>
@@ -64,50 +65,12 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
       |</errorResponse>
     """.stripMargin
 
-  private val MalformedXmlBodyError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>BAD_REQUEST</code>
-      |  <message>Request body does not contain a well-formed XML document.</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val BadRequestErrorXBadgeIdentifierMissingOrInvalid =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>BAD_REQUEST</code>
-      |  <message>X-Badge-Identifier header is missing or invalid</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val InvalidAcceptHeaderError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>ACCEPT_HEADER_INVALID</code>
-      |  <message>The accept header is missing or invalid</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val InvalidContentTypeHeaderError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>UNSUPPORTED_MEDIA_TYPE</code>
-      |  <message>The content type header is missing or invalid</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val InternalServerError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>INTERNAL_SERVER_ERROR</code>
-      |  <message>Internal server error</message>
-      |</errorResponse>
-    """.stripMargin
-
   override protected def beforeAll() {
     startMockServer()
-    stubAuditService()
-    authServiceAuthorizesCSP()
+  }
+
+  override protected def beforeEach() {
+    resetMockServer()
     startMdgWcoDecService()
   }
 
@@ -115,11 +78,109 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
     stopMockServer()
   }
 
+  feature("Declaration API authorises submissions from CSPs and Software Houses with v2.0 accept header") {
+    scenario("An authorised CSP successfully submits a customs clearance declaration") {
+      Given("A CSP wants to submit a valid customs clearance declaration")
+
+      startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
+
+      val request: FakeRequest[AnyContentAsXml] = ValidClearanceRequest.fromCsp.postTo(endpoint)
+
+      And("the CSP is authorised with its privileged application")
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 202 (ACCEPTED) status is received")
+      status(result) shouldBe ACCEPTED
+
+      And("the response body is empty")
+      contentAsString(result) shouldBe 'empty
+
+      And("the request was authorised with AuthService")
+      verifyAuthServiceCalledForCsp()
+
+      And("v2 config was used")
+      verify(1, postRequestedFor(urlEqualTo(CustomsDeclarationsExternalServicesConfig.MdgWcoDecV2ServiceContext)))
+    }
+
+    scenario("An unauthorised CSP is not allowed to submit a customs clearance declaration") {
+      Given("A CSP wants to submit a valid customs clearance declaration")
+      val request: FakeRequest[AnyContentAsXml] = ValidClearanceRequest.fromCsp.postTo(endpoint)
+
+      And("the CSP is unauthorised with its privileged application")
+      authServiceUnauthorisesScopeForCSP()
+      authServiceUnauthorisesCustomsEnrolmentForNonCSP(cspBearerToken)
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 401 (UNAUTHORIZED) status is received")
+      status(result) shouldBe UNAUTHORIZED
+
+      And("the response body is Unauthorised error")
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
+
+      And("the request was authorised with AuthService")
+      verifyAuthServiceCalledForCsp()
+    }
+
+    scenario("A non-CSP successfully submits a declaration on behalf of somebody with Customs enrolment") {
+      Given("A Software House wants to submit a valid customs clearance declaration")
+
+      val request: FakeRequest[AnyContentAsXml] = ValidClearanceRequest.fromNonCsp.postTo(endpoint)
+      startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
+
+      And("declarant is enrolled with Customs having an EORI number")
+      authServiceUnauthorisesScopeForCSP(nonCspBearerToken)
+      authServiceAuthorizesNonCspWithEori()
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 202 (ACCEPTED) status is received")
+      status(result) shouldBe ACCEPTED
+
+      And("the response body is empty")
+      contentAsString(result) shouldBe 'empty
+
+      And("the request was authorised with AuthService")
+      verifyAuthServiceCalledForNonCsp()
+    }
+
+    scenario("A non-CSP is not authorised to submit a declaration on behalf of somebody without Customs enrolment") {
+      Given("A Software House wants to submit a valid customs clearance declaration")
+
+      val request: FakeRequest[AnyContentAsXml] = ValidClearanceRequest.fromNonCsp.postTo(endpoint)
+
+      And("declarant is not enrolled with Customs")
+      authServiceUnauthorisesScopeForCSP(nonCspBearerToken)
+      authServiceUnauthorisesCustomsEnrolmentForNonCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 401 (UNAUTHORIZED) status is received")
+      status(result) shouldBe UNAUTHORIZED
+
+      And("the response body is empty")
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
+
+      And("the request was authorised with AuthService")
+      verifyAuthServiceCalledForNonCsp()
+    }
+  }
+
+  //*** UNHAPPY PATH SPECS ***
+
   feature("The API handles errors as expected") {
 
     scenario("Response status 400 when user submits an xml payload that does not adhere to schema") {
       Given("the API is available")
       val request = InvalidSubmissionRequest.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -138,6 +199,8 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
     scenario("Response status 400 when user submits an xml payload that does not adhere to schema having multiple errors") {
       Given("the API is available")
       val request = InvalidSubmissionRequestWith3Errors.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -156,6 +219,8 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
     scenario("Response status 400 when user submits a malformed xml payload") {
       Given("the API is available")
       val request = MalformedXmlRequest.fromCsp.copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -176,6 +241,8 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
       val request = ValidSubmissionRequest.fromCsp
         .withJsonBody(JsObject(Seq("something" -> JsString("I am a json"))))
         .copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -204,6 +271,9 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
 
       status(resultFuture) shouldBe INTERNAL_SERVER_ERROR
       headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"Internal server error\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InternalServerError)
     }
 
     scenario("Response status 406 when user submits a request without Accept header") {
@@ -245,6 +315,8 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
     scenario("Response status 415 when user submits a request with an invalid Content-Type header") {
       Given("the API is available")
       val request = InvalidContentTypeHeaderSubmissionRequest.copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -263,6 +335,8 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
     scenario("Response status 400 when a CSP user submits a request without a X-Badge-Identifier header") {
       Given("the API is available")
       val request = InvalidSubmissionRequestWithoutXBadgeIdentifier.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -281,6 +355,8 @@ class CustomsDeclarationClearanceUnhappyPathSpec extends AcceptanceTestSpec
     scenario("Response status 400 when a CSP user submits a request with an invalid X-Badge-Identifier header") {
       Given("the API is available")
       val request = InvalidSubmissionRequestWithInvalidXBadgeIdentifier.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
