+++ test/component/CustomsDeclarationSubmissionSpec.scala
@@ -14,22 +14,24 @@
  * limitations under the License.
  */
 
-package acceptance
+package component
 
 import com.github.tomakehurst.wiremock.client.WireMock.{postRequestedFor, urlEqualTo, verify}
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, Matchers, OptionValues}
+import play.api.libs.json.{JsObject, JsString}
 import play.api.mvc._
 import play.api.test.FakeRequest
-import play.api.test.Helpers._
+import play.api.test.Helpers.{status, _}
 import uk.gov.hmrc.customs.declaration.model.{ApiSubscriptionKey, VersionOne, VersionTwo}
 import util.FakeRequests._
+import util.RequestHeaders.X_CONVERSATION_ID_NAME
 import util.TestData._
 import util.externalservices.{ApiSubscriptionFieldsService, AuthService, MdgWcoDecService}
-import util.{CustomsDeclarationsExternalServicesConfig, RequestHeaders}
+import util.{AuditService, CustomsDeclarationsExternalServicesConfig, RequestHeaders}
 
 import scala.concurrent.Future
 
-class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
+class CustomsDeclarationSubmissionSpec extends ComponentTestSpec with AuditService with ExpectedTestResponses
   with Matchers
   with OptionValues
   with BeforeAndAfterAll
@@ -45,11 +47,22 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
   private val apiSubscriptionKeyForXClientIdV2 = apiSubscriptionKeyForXClientIdV1.copy(version = VersionTwo)
 
-  private val UnauthorisedError =
+
+  protected override val BadRequestErrorWith2Errors =
     """<?xml version="1.0" encoding="UTF-8"?>
       |<errorResponse>
-      |  <code>UNAUTHORIZED</code>
-      |  <message>Unauthorised request</message>
+      |  <code>BAD_REQUEST</code>
+      |  <message>Payload is not valid according to schema</message>
+      |  <errors>
+      |     <error>
+      |       <code>xml_validation_error</code>
+      |       <message>cvc-complex-type.3.2.2: Attribute 'foo' is not allowed to appear in element 'Declaration'.</message>
+      |     </error>
+      |     <error>
+      |       <code>xml_validation_error</code>
+      |       <message>cvc-datatype-valid.1.2.1: 'ABC' is not a valid value for 'decimal'.</message>
+      |     </error>
+      |  </errors>
       |</errorResponse>
     """.stripMargin
 
@@ -59,6 +72,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
   override protected def beforeEach() {
     resetMockServer()
+    startMdgWcoDecService()
   }
 
   override protected def afterAll() {
@@ -68,7 +82,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
   feature("Submissions with v1.0 accept header") {
     scenario("An authorised CSP successfully submits a customs declaration") {
       Given("A CSP wants to submit a valid customs declaration")
-      startMdgWcoDecService()
+
       val request: FakeRequest[AnyContentAsXml] = ValidSubmissionRequestWithV1AcceptHeader.fromCsp.postTo(endpoint)
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV1)
 
@@ -96,7 +110,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
   feature("Declaration API authorises submissions from CSPs and Software Houses with v2.0 accept header") {
     scenario("An authorised CSP successfully submits a customs declaration") {
       Given("A CSP wants to submit a valid customs declaration")
-      startMdgWcoDecService()
+
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
 
       val request: FakeRequest[AnyContentAsXml] = ValidSubmissionRequest.fromCsp.postTo(endpoint)
@@ -135,7 +149,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
       status(result) shouldBe UNAUTHORIZED
 
       And("the response body is empty")
-      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedError)
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
 
       And("the request was authorised with AuthService")
       verifyAuthServiceCalledForCsp()
@@ -143,7 +157,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
     scenario("A non-CSP successfully submits a declaration on behalf of somebody with Customs enrolment") {
       Given("A Software House wants to submit a valid customs declaration")
-      startMdgWcoDecService()
+
       val request: FakeRequest[AnyContentAsXml] = ValidSubmissionRequest.fromNonCsp.postTo(endpoint)
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
 
@@ -166,7 +180,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
     scenario("A non-CSP is not authorised to submit a declaration on behalf of somebody without Customs enrolment") {
       Given("A Software House wants to submit a valid customs declaration")
-      startMdgWcoDecService()
+
       val request: FakeRequest[AnyContentAsXml] = ValidSubmissionRequest.fromNonCsp.postTo(endpoint)
 
       And("declarant is not enrolled with Customs")
@@ -180,7 +194,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
       status(result) shouldBe UNAUTHORIZED
 
       And("the response body is empty")
-      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedError)
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
 
       And("the request was authorised with AuthService")
       verifyAuthServiceCalledForNonCsp()
@@ -191,7 +205,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
     scenario("An authorised CSP successfully submits a customs declaration having X-Client-ID request header to v1 api") {
       Given("A CSP wants to submit a valid customs declaration and API Gateway provides X-Client-ID header only")
-      startMdgWcoDecService()
+
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV1)
       val request: FakeRequest[AnyContentAsXml] =
         ValidSubmissionRequestWithXClientIdHeader.withHeaders(RequestHeaders.ACCEPT_HMRC_XML_V1_HEADER).fromCsp.postTo(endpoint)
@@ -218,7 +232,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
     scenario("An authorised CSP successfully submits a customs declaration having X-Client-ID request header to v2 api") {
       Given("A CSP wants to submit a valid customs declaration and API Gateway provides X-Client-ID header only")
-      startMdgWcoDecService()
+
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
       val request: FakeRequest[AnyContentAsXml] = ValidSubmissionRequestWithXClientIdHeader.fromCsp.postTo(endpoint)
 
@@ -244,7 +258,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
     scenario("A non-CSP successfully submits a declaration on behalf of somebody with Customs enrolment having X-Client-ID request header to v1 api") {
       Given("A Software House wants to submit a valid customs declaration and API Gateway provides X-Client-ID header only")
-      startMdgWcoDecService()
+
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV1)
       val request: FakeRequest[AnyContentAsXml] =
         ValidSubmissionRequestWithXClientIdHeader.withHeaders(RequestHeaders.ACCEPT_HMRC_XML_V1_HEADER).fromNonCsp.postTo(endpoint)
@@ -272,7 +286,7 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
 
     scenario("A non-CSP successfully submits a declaration on behalf of somebody with Customs enrolment having X-Client-ID request header to v2 api") {
       Given("A Software House wants to submit a valid customs declaration and API Gateway provides X-Client-ID header only")
-      startMdgWcoDecService()
+
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
       val request: FakeRequest[AnyContentAsXml] = ValidSubmissionRequestWithXClientIdHeader.fromNonCsp.postTo(endpoint)
 
@@ -298,4 +312,225 @@ class CustomsDeclarationSubmissionSpec extends AcceptanceTestSpec
     }
   }
 
+  //*** UNHAPPY PATH SPECS ***
+
+  feature("The API handles errors as expected") {
+
+    scenario("Response status 400 when user submits an xml payload that does not adhere to schema") {
+      Given("the API is available")
+      val request = InvalidSubmissionRequest.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"invalid xml\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestError)
+    }
+
+    scenario("Response status 400 when user submits an xml payload that does not adhere to schema having multiple errors") {
+      Given("the API is available")
+      val request = InvalidSubmissionRequestWith3Errors.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"invalid xml\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestErrorWith2Errors)
+    }
+
+    scenario("Response status 400 when user submits a malformed xml payload") {
+      Given("the API is available")
+      val request = MalformedXmlRequest.fromCsp.copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"malformed xml body\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(MalformedXmlBodyError)
+    }
+
+    scenario("Response status 400 when user submits a non-xml payload") {
+      Given("the API is available")
+      val request = ValidSubmissionRequest.fromCsp
+        .withJsonBody(JsObject(Seq("something" -> JsString("I am a json"))))
+        .copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"malformed xml body\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(MalformedXmlBodyError)
+    }
+
+    scenario("Response status 500 when user submits a request without any client id headers") {
+      Given("the API is available")
+      val request = NoClientIdIdHeaderSubmissionRequest.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 500 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe INTERNAL_SERVER_ERROR
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+
+      And("the response body is an \"Internal server error\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InternalServerError)
+    }
+
+    scenario("Response status 406 when user submits a request without Accept header") {
+      Given("the API is available")
+      val request = NoAcceptHeaderSubmissionRequest.copyFakeRequest(method = POST, uri = endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 406 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe NOT_ACCEPTABLE
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"invalid Accept header\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InvalidAcceptHeaderError)
+    }
+
+    scenario("Response status 406 when user submits a request with an invalid Accept header") {
+      Given("the API is available")
+      val request = InvalidAcceptHeaderSubmissionRequest.copyFakeRequest(method = POST, uri = endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 406 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe NOT_ACCEPTABLE
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"invalid Accept header\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InvalidAcceptHeaderError)
+    }
+
+    scenario("Response status 415 when user submits a request with an invalid Content-Type header") {
+      Given("the API is available")
+      val request = InvalidContentTypeHeaderSubmissionRequest.copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 415 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe UNSUPPORTED_MEDIA_TYPE
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"invalid Content-Type header\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InvalidContentTypeHeaderError)
+    }
+
+    scenario("Response status 400 when a CSP user submits a request without a X-Badge-Identifier header") {
+      Given("the API is available")
+      val request = InvalidSubmissionRequestWithoutXBadgeIdentifier.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"Bad Request\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestErrorXBadgeIdentifierMissingOrInvalid)
+    }
+
+    scenario("Response status 400 when a CSP user submits a request with an invalid X-Badge-Identifier header") {
+      Given("the API is available")
+      val request = InvalidSubmissionRequestWithInvalidXBadgeIdentifier.fromCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceAuthorizesCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then("a response with a 400 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is a \"Bad Request\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(BadRequestErrorXBadgeIdentifierMissingOrInvalid)
+    }
+
+    scenario("Response status 500 when user submits a valid request but downstream call to DMS fails with an HTTP error") {
+
+      Given("the API is available")
+      val request = ValidSubmissionRequest.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      setupMdgWcoDecServiceToReturn(status = NOT_FOUND)
+      val result: Option[Future[Result]] = route(app = app, request)
+
+      Then(s"a response with a 500 status is received")
+      result shouldBe 'defined
+      val resultFuture = result.value
+
+      status(resultFuture) shouldBe INTERNAL_SERVER_ERROR
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+
+      And("the response body is an \"Internal server error\" XML")
+      string2xml(contentAsString(resultFuture)) shouldBe string2xml(InternalServerError)
+    }
+  }
 }
