+++ test/component/DeclarationSchemaSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package acceptance
+package component
 
 import org.scalatest.{Matchers, OptionValues}
 import play.api.mvc._
@@ -27,24 +27,11 @@ import util.externalservices.{ApiSubscriptionFieldsService, AuthService, MdgWcoD
 
 import scala.concurrent.Future
 
-class DeclarationSchemaSpec extends AcceptanceTestSpec
+class DeclarationSchemaSpec extends ComponentTestSpec
   with Matchers with OptionValues with AuthService with MdgWcoDecService with ApiSubscriptionFieldsService with AuditService {
 
   private val endpoint = "/"
 
-  private val BadRequestError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>BAD_REQUEST</code>
-      |  <message>Payload is not valid according to schema</message>
-      |  <errors>
-      |     <error>
-      |       <code>xml_validation_error</code>
-      |       <message>cvc-complex-type.3.2.2: Attribute 'foo' is not allowed to appear in element 'Declaration'.</message>
-      |     </error>
-      |  </errors>
-      |</errorResponse>
-    """.stripMargin
 
   private val apiSubscriptionKeyForXClientIdV1 =
     ApiSubscriptionKey(clientId = clientId, context = "customs%2Fdeclarations", version = VersionOne)
@@ -62,9 +49,9 @@ class DeclarationSchemaSpec extends AcceptanceTestSpec
     stopMockServer()
   }
 
-  feature("The API handles errors as expected") {
+  feature("The API handles Cancellation Specific type code (INV) correctly") {
 
-    scenario("Response status 400 when user submits 13 + INV request") {
+    scenario("Response status 400 when user submits function code 13 + type code INV") {
       Given("the API is available")
       val request = ValidSubmission_13_INV_Request.fromCsp.postTo(endpoint)
 
@@ -80,7 +67,7 @@ class DeclarationSchemaSpec extends AcceptanceTestSpec
     }
 
 
-    scenario("Response status 202 when user submits 13  request") {
+    scenario("Response status 202 when authorised as CSP with privileged application and submits function code 13 in the request") {
       Given("the API is available")
       val request = ValidSubmission_13_Request.fromCsp.postTo(endpoint)
       startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
@@ -94,8 +81,34 @@ class DeclarationSchemaSpec extends AcceptanceTestSpec
       status(result) shouldBe ACCEPTED
     }
 
+    scenario("Response status 400 when authorised as CSP with privileged application and submits type code INV in the request") {
+      Given("the API is available")
+      val request = ValidSubmission_13_INV_Request.fromCsp.postTo(endpoint)
+      startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
+      And("the CSP is authorised with its privileged application")
+      authServiceAuthorizesCSP()
 
-    scenario("Response status 400 when user submits INV request") {
+      When("a POST request with data is sent to the API")
+      val resultFuture: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 400 (BADREQUEST) status is received")
+      status(resultFuture) shouldBe BAD_REQUEST
+      headers(resultFuture).get(X_CONVERSATION_ID_NAME) shouldBe 'defined
+    }
+
+    scenario("Response status 202 when function code 13 is in the request") {
+      Given("the API is available")
+      val request = ValidSubmission_13_Request.fromCsp.postTo(endpoint)
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 202 (ACCEPTED) status is received")
+      status(result) shouldBe ACCEPTED
+    }
+
+
+    scenario("Response status 400 when user submits type code INV in the request") {
       Given("the API is available")
       val request = ValidSubmission_INV_Request.fromCsp.postTo(endpoint)
