+++ test/component/FileUploadSpec.scala
@@ -14,110 +14,148 @@
  * limitations under the License.
  */
 
-package acceptance
+package component
 
-import org.scalatest.{Matchers, OptionValues}
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, Matchers, OptionValues}
 import play.api.libs.json.{JsObject, JsString}
 import play.api.mvc._
-import play.api.test.Helpers._
-import play.mvc.Http.Status.FORBIDDEN
+import play.api.test.FakeRequest
+import play.api.test.Helpers.{status, _}
+import uk.gov.hmrc.customs.declaration.model.{ApiSubscriptionKey, VersionOne, VersionTwo}
+import util.{AuditService, TestData}
 import util.FakeRequests._
 import util.RequestHeaders.X_CONVERSATION_ID_NAME
-import util.TestData.{cspBearerToken, invalidBearerToken}
-import util.externalservices.{AuthService, UpscanInitiateService}
-import util.{AuditService, TestData}
+import util.TestData._
+import util.externalservices.{ApiSubscriptionFieldsService, AuthService, UpscanInitiateService}
 
 import scala.concurrent.Future
 
-class FileUploadUnhappyPathSpec extends AcceptanceTestSpec
-  with Matchers with OptionValues with AuthService with UpscanInitiateService with AuditService {
+class FileUploadSpec extends ComponentTestSpec with ExpectedTestResponses
+  with Matchers
+  with OptionValues
+  with BeforeAndAfterAll
+  with BeforeAndAfterEach
+  with UpscanInitiateService
+  with ApiSubscriptionFieldsService
+  with AuthService
+  with AuditService{
 
   private val endpoint = "/file-upload"
 
-  private val BadRequestSchemaValidationError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>BAD_REQUEST</code>
-      |  <message>Payload is not valid according to schema</message>
-      |  <errors>
-      |     <error>
-      |       <code>xml_validation_error</code>
-      |       <message>cvc-type.3.1.1: Element 'declarationID' is a simple type, so it cannot have attributes, excepting those whose namespace name is identical to 'http://www.w3.org/2001/XMLSchema-instance' and whose [local name] is one of 'type', 'nil', 'schemaLocation' or 'noNamespaceSchemaLocation'. However, the attribute, 'foo' was found.</message>
-      |     </error>
-      |  </errors>
-      |</errorResponse>
-    """.stripMargin
+  private val apiSubscriptionKeyForXClientIdV1 =
+    ApiSubscriptionKey(clientId = clientId, context = "customs%2Fdeclarations", version = VersionOne)
 
-  private val BadRequestErrorXBadgeIdentifierMissingOrInvalid =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>BAD_REQUEST</code>
-      |  <message>X-Badge-Identifier header is missing or invalid</message>
-      |</errorResponse>
-    """.stripMargin
+  private val apiSubscriptionKeyForXClientIdV2 = apiSubscriptionKeyForXClientIdV1.copy(version = VersionTwo)
 
-  private val MalformedXmlBodyError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>BAD_REQUEST</code>
-      |  <message>Request body does not contain a well-formed XML document.</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val InvalidAcceptHeaderError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>ACCEPT_HEADER_INVALID</code>
-      |  <message>The accept header is missing or invalid</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val UnauthorisedRequestError =
-    """<errorResponse>
-      |  <code>UNAUTHORIZED</code>
-      |  <message>Unauthorised request</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val InvalidContentTypeHeaderError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>UNSUPPORTED_MEDIA_TYPE</code>
-      |  <message>The content type header is missing or invalid</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val InternalServerError =
-    """<?xml version="1.0" encoding="UTF-8"?>
-      |<errorResponse>
-      |  <code>INTERNAL_SERVER_ERROR</code>
-      |  <message>Internal server error</message>
-      |</errorResponse>
-    """.stripMargin
-
-  private val ForbiddenError =
-    """<errorResponse>
-      |  <code>FORBIDDEN</code> <message>Not an authorized service</message>
-      |</errorResponse>
-    """.stripMargin
+  private val validUpscanInitiateResponse =
+    """<fileUpload>
+      |<href>https://bucketName.s3.eu-west-2.amazonaws.com</href>
+      |<fields>
+      |<X-Amz-Algorithm>AWS4-HMAC-SHA256</X-Amz-Algorithm>
+      |<X-Amz-Expiration>2018-02-09T12:35:45.297Z</X-Amz-Expiration>
+      |<X-Amz-Signature>xxxx</X-Amz-Signature>
+      |<key>xxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx</key>
+      |<acl>private</acl>
+      |<X-Amz-Credential>ASIAxxxxxxxxx/20180202/eu-west-2/s3/aws4_request</X-Amz-Credential>
+      |<policy>xxxxxxxx==</policy>
+      |</fields>
+      |</fileUpload>""".stripMargin
 
   override protected def beforeAll() {
     startMockServer()
-    stubAuditService()
-    authServiceUnauthorisesScopeForCSP(TestData.nonCspBearerToken)
-    authServiceAuthorizesNonCspWithEori()
-    startUpscanInitiateService()
+  }
+
+  override protected def beforeEach() {
+    resetMockServer()
   }
 
   override protected def afterAll() {
     stopMockServer()
   }
 
+  feature("File upload API authorises submissions from Software Houses with v2.0 accept header") {
+
+    scenario("An unauthorised CSP is not allowed to submit a file upload request") {
+      Given("A CSP wants to submit a valid file upload")
+      val request: FakeRequest[AnyContentAsXml] = ValidFileUploadRequest.fromCsp.postTo(endpoint)
+
+      And("the CSP is unauthorised with its privileged application")
+      authServiceUnauthorisesScopeForCSP()
+      authServiceUnauthorisesCustomsEnrolmentForNonCSP(cspBearerToken)
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 401 (UNAUTHORIZED) status is received")
+      status(result) shouldBe UNAUTHORIZED
+
+      And("the response body is empty")
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
+
+      And("the request was authorised with AuthService")
+      verifyAuthServiceCalledForCsp()
+    }
+
+    scenario("A non-CSP successfully submits a file upload request on behalf of somebody with Customs enrolment") {
+      Given("A Software House wants to submit a valid file upload request")
+      startUpscanInitiateService()
+      val request: FakeRequest[AnyContentAsXml] = ValidFileUploadRequest.fromNonCsp.postTo(endpoint)
+      startApiSubscriptionFieldsService(apiSubscriptionKeyForXClientIdV2)
+
+      And("declarant is enrolled with Customs having an EORI number")
+      authServiceUnauthorisesScopeForCSP(nonCspBearerToken)
+      authServiceAuthorizesNonCspWithEori()
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 200 (OK) status is received")
+      status(result) shouldBe OK
+
+      And("the response body should be correct")
+      string2xml(contentAsString(result)) shouldBe string2xml(validUpscanInitiateResponse)
+
+      And("conversationId is correct")
+      headers(result).get("X-Conversation-ID") shouldBe Some(upscanInitiateReference)
+
+      And("the request was authorised with AuthService")
+      verifyAuthServiceCalledForNonCsp()
+    }
+
+    scenario("A non-CSP is not authorised to submit a file upload request on behalf of somebody without Customs enrolment") {
+      Given("A Software House wants to submit a valid file upload request")
+      startUpscanInitiateService()
+      val request: FakeRequest[AnyContentAsXml] = ValidFileUploadRequest.fromNonCsp.postTo(endpoint)
+
+      And("declarant is not enrolled with Customs")
+      authServiceUnauthorisesScopeForCSP(nonCspBearerToken)
+      authServiceUnauthorisesCustomsEnrolmentForNonCSP()
+
+      When("a POST request with data is sent to the API")
+      val result: Future[Result] = route(app = app, request).value
+
+      Then("a response with a 401 (UNAUTHORIZED) status is received")
+      status(result) shouldBe UNAUTHORIZED
+
+      And("the response body is empty")
+      string2xml(contentAsString(result)) shouldBe string2xml(UnauthorisedRequestError)
+
+      And("the request was authorised with AuthService")
+      verifyAuthServiceCalledForNonCsp()
+    }
+  }
+
+  //*** UNHAPPY PATH SPECS ***
+
   feature("The API handles errors as expected") {
 
     scenario("Response status 400 when user submits an xml payload that does not adhere to schema") {
       Given("the API is available")
       val request = InvalidFileUploadRequest.fromNonCsp.postTo(endpoint)
+      stubAuditService()
+      authServiceUnauthorisesScopeForCSP(TestData.nonCspBearerToken)
+      authServiceAuthorizesNonCspWithEori()
+      startUpscanInitiateService()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -137,6 +175,10 @@ class FileUploadUnhappyPathSpec extends AcceptanceTestSpec
     scenario("Response status 400 when user submits a malformed xml payload") {
       Given("the API is available")
       val request = MalformedXmlRequest.fromNonCsp.copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceUnauthorisesScopeForCSP(TestData.nonCspBearerToken)
+      authServiceAuthorizesNonCspWithEori()
+      startUpscanInitiateService()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -157,6 +199,10 @@ class FileUploadUnhappyPathSpec extends AcceptanceTestSpec
       val request = InvalidFileUploadRequest.fromNonCsp
         .withJsonBody(JsObject(Seq("something" -> JsString("I am a json"))))
         .copyFakeRequest(method = POST, uri = endpoint)
+      stubAuditService()
+      authServiceUnauthorisesScopeForCSP(TestData.nonCspBearerToken)
+      authServiceAuthorizesNonCspWithEori()
+      startUpscanInitiateService()
 
       When("a POST request with data is sent to the API")
       val result: Option[Future[Result]] = route(app = app, request)
@@ -308,5 +354,4 @@ class FileUploadUnhappyPathSpec extends AcceptanceTestSpec
       string2xml(contentAsString(resultFuture)) shouldBe string2xml(InternalServerError)
     }
   }
-
 }
