+++ test/component/definitionSpecs.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package acceptance
+package component
 
 import org.scalatest.Matchers
 import play.api.Application
@@ -26,7 +26,7 @@ import play.api.test.Helpers._
 
 import scala.concurrent.Future
 
-class DefinitionSpecWithAllVersionsEnabledByDefault extends AcceptanceTestSpec with Matchers {
+class DefinitionSpecWithAllVersionsEnabledByDefault extends ComponentTestSpec with Matchers {
 
   override implicit lazy val app: Application = new GuiceApplicationBuilder().configure(Map(
     "api.access.version-2.0.whitelistedApplicationIds.0" -> "someId-1",
@@ -120,7 +120,7 @@ class DefinitionSpecWithAllVersionsEnabledByDefault extends AcceptanceTestSpec w
   }
 }
 
-class DefinitionSpecWithVersion2Disabled extends AcceptanceTestSpec with Matchers {
+class DefinitionSpecWithVersion2Disabled extends ComponentTestSpec with Matchers {
 
   override implicit lazy val app: Application = new GuiceApplicationBuilder().configure(Map(
     "api.access.version-2.0.enabled" -> false
